+++ lib/src/televerse/context/inline_query.dart
@@ -4,12 +4,11 @@ part of televerse.context;
 /// with [InlineQueryMixin]. It contains the inline query and the update. The context will be passed to the inline query handler, and once you have the context you can simply use the methods to reply or answer the query.
 class InlineQueryContext extends Context with InlineQueryMixin {
   /// The incoming inline query.
-  InlineQuery query;
+  InlineQuery get query => update.inlineQuery!;
 
   /// Constructs a new [InlineQueryContext].
   InlineQueryContext(
-    RawAPI t,
-    this.query, {
+    RawAPI t, {
     required Update update,
   }) : super(t, update: update);