+++ lib/src/types/aliases.dart
@@ -21,3 +21,37 @@ typedef InlineQueryHandler = FutureOr<void> Function(InlineQueryContext ctx);
 /// This is used to define a callback function for almost all the methods that listen to chat member updates.
 typedef ChatMemberUpdatedHandler = FutureOr<void> Function(
     ChatMemberUpdatedContext ctx);
+
+/// [PollHandler] is a type alias for a function that takes a [PollContext] as parameter and returns a [FutureOr] of void.
+///
+/// This is used to define a callback function for almost all the methods that listen to polls.
+typedef PollHandler = FutureOr<void> Function(PollContext ctx);
+
+/// [PollAnswerHandler] is a type alias for a function that takes a [PollAnswerContext] as parameter and returns a [FutureOr] of void.
+///
+/// This is used to define a callback function for almost all the methods that listen to poll answers.
+typedef PollAnswerHandler = FutureOr<void> Function(PollAnswerContext ctx);
+
+/// [ChosenInlineResultHandler] is a type alias for a function that takes a [ChosenInlineResultContext] as parameter and returns a [FutureOr] of void.
+///
+/// This is used to define a callback function for almost all the methods that listen to chosen inline results.
+typedef ChosenInlineResultHandler = FutureOr<void> Function(
+    ChosenInlineResultContext ctx);
+
+/// [ChatJoinRequestHandler] is a type alias for a function that takes a [ChatJoinRequestContext] as parameter and returns a [FutureOr] of void.
+///
+/// This is used to define a callback function for almost all the methods that listen to chat join requests.
+typedef ChatJoinRequestHandler = FutureOr<void> Function(
+    ChatJoinRequestContext ctx);
+
+/// [ShippingQueryHandler] is a type alias for a function that takes a [ShippingQueryContext] as parameter and returns a [FutureOr] of void.
+///
+/// This is used to define a callback function for almost all the methods that listen to shipping queries.
+typedef ShippingQueryHandler = FutureOr<void> Function(
+    ShippingQueryContext ctx);
+
+/// [PreCheckoutQueryHandler] is a type alias for a function that takes a [PreCheckoutQueryContext] as parameter and returns a [FutureOr] of void.
+///
+/// This is used to define a callback function for almost all the methods that listen to pre checkout queries.
+typedef PreCheckoutQueryHandler = FutureOr<void> Function(
+    PreCheckoutQueryContext ctx);