+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## 1.8.2
+- Added `PollContext`, `PollAnswerContext` classes to represent the context of a `poll` and `poll_answer` updates.
+- Added `Televerse.poll` method to listen to `poll` updates.
+- Added `Televerse.pollAnswer` method to listen to `poll_answer` updates.
+- Now most of the Televerse methods return a `StreamSubscription` instead of `void`.
+- This is to make it easier to cancel the subscription and increase the performance.
+- Added `ChosenInlineResultContext` class to represent the context of a `chosen_inline_result` update and `Televerse.chosenInlineResult` method to listen to `chosen_inline_result` updates.
+- Added `ChatJoinRequestContext` class to represent the context of a `chat_join_request` update and `Televerse.chatJoinRequest` method to listen to `chat_join_request` updates.
+- Added `ShippingQueryContext` class to represent the context of a `shipping_query` update and `Televerse.shippingQuery` method to listen to `shipping_query` updates.
+- Added `PreCheckoutQueryContext` class to represent the context of a `pre_checkout_query` update and `Televerse.preCheckoutQuery` method to listen to `pre_checkout_query` updates.
 ## 1.8.1
 - Added `ChatMemberUpdatedContext` class to represent the context of a `chat_member` and `my_chat_member` updates.
 - Added `Televerse.chatMember` and `Televerse.myChatMember` methods to listen to `chat_member` and `my_chat_member` updates.